+++ .gitignore
@@ -1,3 +1,1 @@
 /tqs-gohouse/target/
-/tqs-gohouse/nbproject/
-/tqs-gohouse/.fr-xoFbYg/tqs-gohouse-master/tqs-gohouse/nbproject/
