--- tqs-gohouse/src/main/java/dbClasses/Room.java
@@ -1,117 +1,0 @@
-package dbClasses;
-
-import java.io.Serializable;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Room implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    
-    @Column(nullable = false)
-    private int rent;
-
-    private String description;
-    
-    @ManyToOne
-    @JoinColumn(nullable=false)
-    private Property property;
-    
-    @ManyToOne
-    @JoinColumn()
-    private PlatformUser renter;
-
-    private boolean occupied;
-
-    public Room() {
-    }
-
-    public Room(int rent, University owner) {
-        this.rent = rent;
-        this.renter = null;
-        this.occupied = false;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Property getProperty() {
-        return property;
-    }
-
-    public void setProperty(Property property) {
-        this.property = property;
-    }
-    
-    public int getRent() {
-        return rent;
-    }
-
-    public void setRent(int rent) {
-        this.rent = rent;
-    }
-
-    public PlatformUser getRenter() {
-        return renter;
-    }
-
-    public void setRenter(PlatformUser renter) {
-        this.renter = renter;
-    }
-
-    public boolean isOccupied() {
-        return occupied;
-    }
-
-    public void setOccupied(boolean occupied) {
-        this.occupied = occupied;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    @Override
-    public int hashCode() {
-        int hash = 0;
-        hash += (id != null ? id.hashCode() : 0);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof Room)) {
-            return false;
-        }
-        Room other = (Room) object;
-        if ((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id))) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "dbClasses.Room[ id=" + id + " ]";
-    }
-    
-}
